requirements.txt
numpy
pandas
sentence-transformers
faiss-cpu
transformers
pytest
